Version 5 - Baseline
This is the code that more or less is the basis of our model. While it
doesn’t have much of the customized settings we would like to test
with, this model can be used to solve any question we would like to ask
in the analysis of our model.